+++ lib/runners/processor/pylint.rb
@@ -71,6 +71,11 @@ module Runners
     def run_analyzer
       files = analyzed_files
 
+      if files.empty?
+        trace_writer.message "No files found."
+        return Results::Success.new(guid: guid, analyzer: analyzer)
+      end
+
       trace_writer.message "Analyzing #{files.size} file(s)..."
 
       stdout, stderr = capture3!(